Create README.md
# Poupulation-Data(Sub-Saharan-Africa).
Sub-Saharan Africa is, geographically, the area and regions of the continent of Africa that lie south of the Sahara. These include Central, East, Southern, and West Africa. In this SQL exploration, we find insights into the growth of the African population.
The exploration also seeks to compare the birthrate and death rates and compare the female to male populations.